Import of MARCEL/any-feature-0.02 from CPAN.
gitpan-cpan-distribution: any-feature
gitpan-cpan-version:      0.02
gitpan-cpan-path:         MARCEL/any-feature-0.02.tar.gz
gitpan-cpan-author:       MARCEL
gitpan-cpan-maturity:     released

+++ xt/author/05_kwalitee.t
@@ -1,9 +1,9 @@
 #!perl
-
 use strict;
 use warnings;
 use FindBin '$Bin';
 use Test::Kwalitee;
-
 my $file = "$Bin/../../Debian_CPANTS.txt";
-unlink $file or die "can't unlink $file: $!\n";
+if (-e $file) {
+    unlink $file or die "can't unlink $file: $!\n";
+}
